+++ network/api.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	"github.com/Scalingo/go-plugins-helpers/sdk"
+	"github.com/Scalingo/go-plugins-helpers/v2/sdk"
 	"github.com/Scalingo/go-utils/logger"
 )